Rename run -> tickAnimator
